requirements.txt
streamlit
asana
fuzzywuzzy 
python-Levenshtein 
python-dotenv